User dongwon

+++ src/main/java/com/example/newsfeed/domain/user/service/UserService.java
@@ -42,8 +42,11 @@ public class UserService {
 
     // 회원탈퇴
     @Transactional
-    public void withdrawal(Long id) {
+    public void withdrawal(Long id, String password) {
         User user = findByIdOrElseThrow(id);
+        if (!passwordEncoder.matches(password, user.getPassword())) {
+            throw new PasswordMismatchException("비밀번호가 올바르지 않습니다.");
+        }
         userRepository.delete(user);
     }
 
@@ -72,9 +75,13 @@ public class UserService {
 
     @Transactional
     public void updatePassword(Long id, String oldPassword, String newPassword) {
+
         User user = findByIdOrElseThrow(id);
         if (!passwordEncoder.matches(oldPassword, user.getPassword())) {
             throw new PasswordMismatchException("비밀번호가 올바르지 않습니다.");
+        }
+        if (passwordEncoder.matches(newPassword, user.getPassword())) {
+            throw new IllegalArgumentException("현재 비밀번호와 동일한 비밀번호로는 변경할 수 없습니다.");
         }
 
         user.setEncodedPassword(passwordEncoder.encode(newPassword));
